[Feat] :  컴포넌트 분리, 판매상품 데이터 이미지 렌더링 - (서린 03:06)
