+++ src/components/01_atoms/Divider/Divider.tsx
@@ -1,0 +1,3 @@
+export const Divider = () => {
+  return <div className="border-b border-[#eee] my-20"></div>;
+};
